Add caching logic
This cache is made to be light on resources and store mangas on an
as-needed basis.

Since we only care about the user's manga titles, store those entries
in the cache if there's a hit from a source.

On subsequent fetches, the chapter numbers from the cache and updates
are compared to see if an update notification should be sent.

There is also a TTL on each cache entry which is either picked up and
cleaned by the update function or a background clean cache task.

Signed-off-by: kingbri <bdashore3@gmail.com>