MangaLife: Fix times
When subtracting today's date and the mangalife timestamp, a negative
number sometimes came out of the function.

This shouldn't make sense since the released manga is from the future?

This error was happening because MangaLife's timestamps are 1 hour
behind what Date.parse() gives in javascript.

Remove 1 hour (equivalent in ms) from the mangalife timestamp on every
parse to get the correct times.

Signed-off-by: kingbri <bdashore3@gmail.com>