Index: Split handling of updates
Every step of the update procedure is now its individual function.

In addition, disable MangaDex as the source won't have updates due to
its read-only state.

Add the predecessor code for setting an interval to dispatch updates.
However, we won't be using this yet as the code is still being tested.

Signed-off-by: kingbri <bdashore3@gmail.com>