Cache: Fix overall logic
ES6 maps are very weird when comparing object keys. Instead of comparing
each value of the object, the memory addresses of the object are compared
to look up the value.

This is not optimal for our situation since we have the TTL as the map
value and we are looking for the presence of a value with a MangaPacket key.

Thankfully, our MangaPackets are small and can easily be stringified into
JSON strings. The cache now consists of JSON strings instead of MangaPacket
objects.

The other problem was storing the elapsed time inside a cache entry.
However, elapsed time is only used within the sources, so there's no
need to store elapsed time in a MangaPacket.

Remove the elapsed time value from the MangaPacket object and its
addition from all sources.

Signed-off-by: kingbri <bdashore3@gmail.com>