Index: Strip unnecessary characters from manga title
The foundation for a future commit which encodes the user JSON file
into a lowercase string with no punctuation and proper spacing.

This change adds this functionality when comparing titles between
the source and user configuration.

Since all manga titles are extremely unique, removing punctuation
and capitalization isn't a huge issue.

To fix captalization issues with the manga title on notification push,
use the filter function to set the manga title to the one from
the source as the user title.

Signed-off-by: kingbri <bdashore3@gmail.com>