Cache: use node-cache
If two different users have the same MangaPacket in updates, the cache
will hit for the second user and that person's update will not be sent.

To fix this, add a user array for the value field of the cache. However,
this means that the TTL is obsolete due to using a Map for the previous cache.

Switch to using node-cache as a drop-in replacement with its own TTL
field. This removes the flushCache function and separate interval.

Signed-off-by: kingbri <bdashore3@gmail.com>