Fix manual setup exceptions
Sources end with a colon by default. However, the manual parser
was filtering if a string included a colon and not checking if the
end of a string specifically has a colon

Titles were also not being converted to a readable format. Even
though this is machine JSON, readability is required for debugging
purposes.

Remove all apostrophes using the replaceAll method and trim
the manga title before returning.

Signed-off-by: kingbri <bdashore3@gmail.com>