Actions: Add server package workflow
This will be used to package the server for deploys.

Since the server won't run without the necessary modules post
compilation, opt to package the binary as standalone using pkg.

No sensitive code is stored within the executable itself, so this
is fine for running MangaUpdates on a cloud server.

Signed-off-by: kingbri <bdashore3@gmail.com>

Fix 1

Fix 2

Fix 3