Fix caching with images
Manga covers can change their IDs on the fly due to using a different
image for a cover.

Keeping this volatile parameter in the cache will cause cache misses
and then send notifications more than once.

The only way to fix this and not affect the original MangaPacket is
to clone the object and delete the image parameter from the cloned
object.

Therefore, when the object is stringified and checked against the cache,
the result should be accurate.

Also make the image parameter optional since the URL could be undefined
and make MangaDex return an undefined image parameter for now.

Signed-off-by: kingbri <bdashore3@gmail.com>