Services: Update chapter renaming
Move chapter renaming into a utils function since this will be used
in all services.

This function tests if a chapter is a standalone number and parses
that for the description of a notification push.

Ifttt requires 3 ingredients per webhook push. With the addition
of the image URL, ifttt has now been changed to create the description
within the code rather than through the applet.

With the old chapter renaming function, we tested if the string contained
'ch' within it. However, this isn't practical as we are checking
if the chapter is a standalone number or a string.

Therefore, parse float is used to check if the number is a decimal
or integer and handle the chapter string from there.

Signed-off-by: kingbri <bdashore3@gmail.com>