Add update push logic for databases
With a database, a constant connection is maintained. The best option
is to minimize the amount of database calls. This is why sql calls
are staggered within different functions of the update notifier.

Also use Promise.allSettled to run all asynchronous tasks in parallel
for each user. This increases efficiency of each update check since
every user is checked at the same time instead of in series.

Add some types for a generic service that is used by the database.

Signed-off-by: kingbri <bdashore3@gmail.com>