Updater: Check for source inclusion vs equality
Checking if the source names are equal will not work if the names
are changed on the user/server side.

For example, MangaFox is mangafox2 in paperback backups and bakaupdates
will include the original scanlator on notification.

Therefore, checking for source inclusion is more flexible than using
equality operators.

NOTE: This commit may fail depending on the user's JSON sources.

Signed-off-by: kingbri <bdashore3@gmail.com>