MangaDex: Cleanup and type additions
The previous MD URL to fetch updated chapters was malformed. Fix
this and update the fetch function to reflect the new behavior.

Also add types for MangaDex responses which allow us to accurately
develop with these results in mind.

Covers are not added in this commit since that requires a different
API route.

Signed-off-by: kingbri <bdashore3@gmail.com>