Search: Added Index.ts
Added an index file to search through all the sources and returns an array of search packet. (Implementation from sourceUpdates/index)

The index file also sanitises the userinput so no string checking is required for searching prior to runing the function.
Left remove chars and trims in individual source file ( in a case where we want to search to a specific source and not through index file)