Search: searchTest
Added another function that calls on the index file in SourceSearch to return an array of searchPackets without any print statements.

The testSearches function still exists.

The file takes an additional argument in the 3rd position which defines which of the two functions to execute.

Updated console arg example on top of the file.